eMANUEL

+++ src/main/java/com/betha/statustce/statustce/model/UnidadeAtendimento.java
@@ -20,20 +20,20 @@ public class UnidadeAtendimento {
 
     @Id
     @Column(name="id")
-    private Integer id;
+    private Long id;
     @Column(name="nome")
     private String nome;
 
-    public UnidadeAtendimento(Integer id, String nome) {
+    public UnidadeAtendimento(Long id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
